Module darwinia_service::chain_spec

Polkadot chain configurations.

Structs

Extensions
Node ChainSpec extensions.

Functions

crab_config

crab_development_config
Crab development config (single validator Alice)

crab_genesis_builder_config
Crab config.

crab_local_testnet_config
Crab local testnet config (multivalidator Alice + Bob)

crab_testnet_genesis
Helper function to create Crab GenesisConfig for testing

get_account_id_from_seed
Helper function to generate an account ID from seed

get_authority_keys_from_seed
Helper function to generate stash, controller and session key from seed

get_from_seed
Helper function to generate a crypto pair from seed

Type Definitions

CrabChainSpec
The ChainSpec parametrised for Crab runtime.