Struct darwinia_service::Configuration

pub struct Configuration {
    pub impl_name: &'static str,
    pub impl_version: &'static str,
    pub impl_commit: &'static str,
    pub roles: Roles,
    pub task_executor: Option<Arc<dyn Fn(Pin<Box<dyn Future<Output = ()> + 'static + Send>>) + 'static + Send + Sync>>,
    pub transaction_pool: Options,
    pub network: NetworkConfiguration,
    pub config_dir: Option<PathBuf>,
    pub keystore: KeystoreConfig,
    pub database: Option<DatabaseConfig>,
    pub state_cache_size: usize,
    pub state_cache_child_ratio: Option<usize>,
    pub pruning: PruningMode,
    pub chain_spec: Option<Box<dyn ChainSpec + 'static>>,
    pub name: String,
    pub wasm_method: WasmExecutionMethod,
    pub execution_strategies: ExecutionStrategies,
    pub rpc_http: Option<SocketAddr>,
    pub rpc_ws: Option<SocketAddr>,
    pub rpc_ws_max_connections: Option<usize>,
    pub rpc_cors: Option<Vec<String>>,
    pub prometheus_config: Option<PrometheusConfig>,
    pub telemetry_endpoints: Option<TelemetryEndpoints>,
    pub telemetry_external_transport: Option<ExtTransport>,
    pub default_heap_pages: Option<u64>,
    pub offchain_worker: bool,
    pub sentry_mode: bool,
    pub force_authoring: bool,
    pub disable_grandpa: bool,
    pub dev_key_seed: Option<String>,
    pub tracing_targets: Option<String>,
    pub tracing_receiver: TracingReceiver,
    pub max_runtime_instances: usize,
}

Service configuration.

Fields
impl_name: &'static str

Implementation name

impl_version: &'static str

Implementation version

impl_commit: &'static str

Git commit if any.

roles: Roles

Node roles.

task_executor: Option<Arc<dyn Fn(Pin<Box<dyn Future<Output = ()> + 'static + Send>>) + 'static + Send + Sync>>

How to spawn background tasks. Mandatory, otherwise creating a Service will error.

transaction_pool: Options

Extrinsic pool configuration.

network: NetworkConfiguration

Network configuration.

config_dir: Option<PathBuf>

Path to the base configuration directory.

keystore: KeystoreConfig

Configuration for the keystore.

database: Option<DatabaseConfig>

Configuration for the database.

state_cache_size: usize

Size of internal state cache in Bytes

state_cache_child_ratio: Option<usize>

Size in percent of cache size dedicated to child tries

pruning: PruningMode

Pruning settings.

chain_spec: Option<Box<dyn ChainSpec + 'static>>

Chain configuration.

name: String

Node name.

wasm_method: WasmExecutionMethod

Wasm execution method.

execution_strategies: ExecutionStrategies

Execution strategies.

rpc_http: Option<SocketAddr>

RPC over HTTP binding address. None if disabled.

rpc_ws: Option<SocketAddr>

RPC over Websockets binding address. None if disabled.

rpc_ws_max_connections: Option<usize>

Maximum number of connections for WebSockets RPC server. None if default.

rpc_cors: Option<Vec<String>>

CORS settings for HTTP & WS servers. None if all origins are allowed.

prometheus_config: Option<PrometheusConfig>

Prometheus endpoint configuration. None if disabled.

telemetry_endpoints: Option<TelemetryEndpoints>

Telemetry service URL. None if disabled.

telemetry_external_transport: Option<ExtTransport>

External WASM transport for the telemetry. If Some, when connection to a telemetry endpoint, this transport will be tried in priority before all others.

default_heap_pages: Option<u64>

The default number of 64KB pages to allocate for Wasm execution

offchain_worker: bool

Should offchain workers be executed.

sentry_mode: bool

Sentry mode is enabled, the node's role is AUTHORITY but it should not actively participate in consensus (i.e. no keystores should be passed to consensus modules).

force_authoring: bool

Enable authoring even when offline.

disable_grandpa: bool

Disable GRANDPA when running in validator mode

dev_key_seed: Option<String>

Development key seed.

When running in development mode, the seed will be used to generate authority keys by the keystore.

Should only be set when node is running development mode.

tracing_targets: Option<String>

Tracing targets

tracing_receiver: TracingReceiver

Tracing receiver

max_runtime_instances: usize

The size of the instances cache.

The default value is 8.

Methods

impl Configuration

pub fn from_version(version: &VersionInfo) -> Configuration

Create a default config using VersionInfo

pub fn full_version(&self) -> String

Returns full version string of this configuration.

pub fn client_id(&self) -> String

Implementation id and version.

pub fn in_chain_config_dir(&self, sub: &str) -> Option<PathBuf>

Generate a PathBuf to sub in the chain configuration directory if given

pub fn expect_chain_spec(&self) -> &dyn ChainSpec

Return a reference to the ChainSpec of this Configuration.

Panics

This method panic if the chain_spec is None

pub fn expect_database(&self) -> &DatabaseConfig

Return a reference to the DatabaseConfig of this Configuration.

Panics

This method panic if the database is None

pub fn display_role(&self) -> String

Returns a string displaying the node role, special casing the sentry mode (returning SENTRY), since the node technically has an AUTHORITY role but doesn't participate.

pub fn use_in_memory_keystore(&mut self) -> Result<(), String>

Use in memory keystore config when it is not required at all.

This function returns an error if the keystore is already set to something different than KeystoreConfig::None.